Create SendMail.cs
implement a mail user agent that sends mail to other users
using gmail. I programmed the SMTP interaction between the MUA and the gmail SMTP server. The client provides a graphical user interface containing fields for entering the sender and recipient addresses, the subject of the message and the message itself.